# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StoreKit API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /collections/{slug}:
    get:
      operationId: getCollectionBySlug
      tags: [collections]
      summary: Get collection by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/auth/login:
    post:
      operationId: adminLogin
      tags: [admin-auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/auth/check:
    get:
      operationId: adminAuthCheck
      tags: [admin-auth]
      summary: Check admin auth status
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/collections:
    get:
      operationId: adminListCollections
      tags: [admin-collections]
      summary: Admin - list all collections
      responses:
        "200":
          description: Collections list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"
    post:
      operationId: adminCreateCollection
      tags: [admin-collections]
      summary: Admin - create collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollectionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
  /admin/collections/{id}:
    patch:
      operationId: adminUpdateCollection
      tags: [admin-collections]
      summary: Admin - update collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCollectionBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
    delete:
      operationId: adminDeleteCollection
      tags: [admin-collections]
      summary: Admin - delete collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin-orders]
      summary: Admin - list all orders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderListResponse"
  /admin/orders/{id}:
    get:
      operationId: adminGetOrder
      tags: [admin-orders]
      summary: Admin - get order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: adminUpdateOrder
      tags: [admin-orders]
      summary: Admin - update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /admin/testimonials:
    get:
      operationId: adminListTestimonials
      tags: [admin-content]
      summary: Admin - list testimonials
      responses:
        "200":
          description: Testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: adminCreateTestimonial
      tags: [admin-content]
      summary: Admin - create testimonial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestimonialBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
  /admin/testimonials/{id}:
    patch:
      operationId: adminUpdateTestimonial
      tags: [admin-content]
      summary: Admin - update testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTestimonialBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
    delete:
      operationId: adminDeleteTestimonial
      tags: [admin-content]
      summary: Admin - delete testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/banners:
    get:
      operationId: adminListBanners
      tags: [admin-content]
      summary: Admin - list banners
      responses:
        "200":
          description: Banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: adminCreateBanner
      tags: [admin-content]
      summary: Admin - create banner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBannerBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /admin/banners/{id}:
    patch:
      operationId: adminUpdateBanner
      tags: [admin-content]
      summary: Admin - update banner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBannerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: adminDeleteBanner
      tags: [admin-content]
      summary: Admin - delete banner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Collection:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
        productCount:
          type: integer
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - slug
        - name
        - description
        - isFeatured
        - sortOrder
        - productCount
        - createdAt
        - updatedAt
    ShippingAddress:
      type: object
      properties:
        fullName:
          type: string
        line1:
          type: string
        line2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
        phone:
          type: string
      required: [fullName, line1, city, state, postalCode, country, phone]
    OrderItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        variantId:
          type: string
        productName:
          type: string
        variantLabel:
          type: string
        imageUrl:
          type: string
        price:
          type: integer
        quantity:
          type: integer
        total:
          type: integer
      required:
        - id
        - productId
        - variantId
        - productName
        - variantLabel
        - imageUrl
        - price
        - quantity
        - total
    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        userId:
          type: string
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, refunded]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        subtotal:
          type: integer
        shippingCost:
          type: integer
        tax:
          type: integer
        total:
          type: integer
        stripePaymentIntentId:
          type: ["string", "null"]
        trackingNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - orderNumber
        - userId
        - status
        - items
        - shippingAddress
        - subtotal
        - shippingCost
        - tax
        - total
        - createdAt
        - updatedAt
    AdminOrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required: [orders, total, page, pageSize, totalPages]
    CreateOrderBody:
      type: object
      properties:
        userId:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              variantId:
                type: string
              quantity:
                type: integer
            required: [variantId, quantity]
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        stripePaymentIntentId:
          type: string
      required: [userId, items, shippingAddress, stripePaymentIntentId]
    UpdateOrderBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, refunded]
        trackingNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...
    CreateCollectionBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
      required: [name, slug, description]
    UpdateCollectionBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
        metaTitle:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
...
    Testimonial:
      type: object
      properties:
        id:
          type: string
        authorName:
          type: string
        authorLocation:
          type: string
        text:
          type: string
        rating:
          type: integer
        isVisible:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
      required: [id, authorName, authorLocation, text, rating, isVisible, sortOrder, createdAt]
    CreateTestimonialBody:
      type: object
      properties:
        authorName:
          type: string
        authorLocation:
          type: string
        text:
          type: string
        rating:
          type: integer
        isVisible:
          type: boolean
        sortOrder:
          type: integer
      required: [authorName, text, rating]
    UpdateTestimonialBody:
      type: object
      properties:
        authorName:
          type: string
        authorLocation:
          type: string
        text:
          type: string
        rating:
          type: integer
        isVisible:
          type: boolean
        sortOrder:
          type: integer
    Banner:
      type: object
      properties:
        id:
          type: string
        imageUrl:
          type: string
        heading:
          type: string
        subheading:
          type: string
        ctaText:
          type: string
        ctaUrl:
          type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, imageUrl, heading, subheading, ctaText, ctaUrl, sortOrder, isActive, createdAt]
    CreateBannerBody:
      type: object
      properties:
        imageUrl:
          type: string
        heading:
          type: string
        subheading:
          type: string
        ctaText:
          type: string
        ctaUrl:
          type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean
      required: [imageUrl]
    UpdateBannerBody:
      type: object
      properties:
        imageUrl:
          type: string
        heading:
          type: string
        subheading:
          type: string
        ctaText:
          type: string
        ctaUrl:
          type: string
        sortOrder:
          type: integer
        isActive:
          type: boolean
    AdminLoginBody:
      type: object
      properties:
        password:
          type: string
      required: [password]
...