# lib/api-zod/src/generated/types/shippingAddress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StoreKit API
 * OpenAPI spec version: 0.1.0
 */

export interface ShippingAddress {
  fullName: string;
  line1: string;
  /** @nullable */
  line2?: string | null;
  city: string;
  state: string;
  postalCode: string;
  country: string;
  phone: string;
}