# lib/api-zod/src/generated/types/updateCollectionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StoreKit API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateCollectionBody {
  name?: string;
  slug?: string;
  description?: string;
  /** @nullable */
  imageUrl?: string | null;
  isFeatured?: boolean;
  sortOrder?: number;
  /** @nullable */
  metaTitle?: string | null;
  /** @nullable */
  metaDescription?: string | null;
}